just developing
gnjfg

+++ app/Http/Controllers/ProductsController.php
@@ -4,7 +4,10 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Product;
+use App\Category;
 use App\Sub_category;
+use App\Sport;
+use App\Brands;
 
 class ProductsController extends Controller
 {
@@ -13,16 +16,13 @@ class ProductsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create($userId, $categoryId, $subCategoryId)
     {
-		if (!$categories = Category::all()) {
+		if (!Auth::check() || Auth::user()->role!="root") {
 			abort(404);
 		}
 
-		if (!$sub_categories = Sub_category::all()) {
-			abort(404);
-		}
-        //llamar a vista de crear producto con arrays de categories y subcategpries
+		return view('partials.admin.formProduct', array('userId'=>$userId, 'categoryId'=>$categoryId, 'subCategoryId'=>$subCategoryId));
     }
 
     /**
@@ -31,8 +31,27 @@ class ProductsController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Product $request)
+    public function store(Request $request, $userId, $categoryId, $subCategoryId)
     {
+		if (!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
+		$validateData = $request->validate([
+			'name' => 'required|max:100',
+    		'description' => 'required',
+    		'shortDescription' => 'required|max:50',
+    		'basePrice' => 'required|numeric',
+    		'typeSupplement' => 'max:50',
+    		'volume' => 'numeric',
+    		'shifts' => 'integer',
+    		'wheelsAmount' => 'integer',
+    		'weight' => 'numeric',
+    		'subCategoryId' => 'required|integer',
+    		'brandId' => 'required|integer',
+    		'sportId' => 'required|integer'
+	  	]);
+
 		$product = new Product();
         $product->name = $request->input('name');
 		$product->description = $request->input('description');
@@ -44,11 +63,12 @@ class ProductsController extends Controller
 		$product->shifts = $request->input('shifts');
 		$product->wheelsAmount = $request->input('wheelsAmount');
 		$product->weight = $request->input('weight');
-		$product->subCategory = $request->input('subCategory');
-		$product->brandId = $request->input('brandId');
-		$product->sportId = $request->input('sportId');
+		$product->subCategory = $subCategoryId;
+		$product->brandId = $brandId;
+		$product->sportId = $sportId;
 
 		$product->save();
+		redirect("user/$userId/categories/$categoryId/sub_categories/$subCategoryId");
     }
 
     /**
@@ -57,8 +77,12 @@ class ProductsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($productId)
+    public function show($userId, $categoryId, $subCategoryId, $productId)
     {
+		if (!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
 		if (!$product = Product::find($productId)) {
 			abort(404);
 		}
@@ -72,12 +96,17 @@ class ProductsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($productId)
+    public function edit($userId, $categoryId, $subCategoryId, $productId)
     {
+		if (!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
 		if (!$product = Sub_category::find($productId)) {
 			abort(404);
 		}
-		//llamar a vista de editar producto
+
+		return view('partials.admin.formProduct', array('userId'=>$userId, 'categoryId'=>$categoryId, 'subCategoryId'=>$subCategoryId, 'product'=>$product));
     }
 
     /**
@@ -87,12 +116,32 @@ class ProductsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Product $request, $productId)
+    public function update(Request $request, $userId, $categoryId, $subCategoryId, $productId)
     {
+		if (!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
 		if (!$product = Product::find($productId)) {
 			abort(404);
 		}
 
+		$validateData = $request->validate([
+			'name' => 'required|max:100',
+    		'description' => 'required',
+    		'shortDescription' => 'required|max:50',
+    		'basePrice' => 'required|numeric',
+    		'typeSupplement' => 'max:50',
+    		'volume' => 'numeric',
+    		'shifts' => 'integer',
+    		'wheelsAmount' => 'integer',
+    		'weight' => 'numeric',
+    		'subCategoryId' => 'required|integer',
+    		'brandId' => 'required|integer',
+    		'sportId' => 'required|integer'
+	  	]);
+
+        $product->name = $request->input('name');
 		$product->description = $request->input('description');
 		$product->shortDescription = $request->input('shortDescription');
 		$product->basePrice = $request->input('basePrice');
@@ -102,11 +151,12 @@ class ProductsController extends Controller
 		$product->shifts = $request->input('shifts');
 		$product->wheelsAmount = $request->input('wheelsAmount');
 		$product->weight = $request->input('weight');
-		$product->subCategory = $request->input('subCategory');
-		$product->brandId = $request->input('brandId');
-		$product->sportId = $request->input('sportId');
+		$product->subCategory = $subCategoryId;
+		$product->brandId = $brandId;
+		$product->sportId = $sportId;
 
 		$product->save();
+		redirect("user/$userId/categories/$categoryId/sub_categories/$subCategoryId");
     }
 
     /**
@@ -115,12 +165,24 @@ class ProductsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($productId)
+    public function destroy($userId, $categoryId, $subCategoryId, $productId)
     {
+		if (!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
 		if (!$product = Product::find($productId)) {
 			abort(404);
 		}
 
+		$images = Image::where('productId', $productId)->get();
+		if (count($images)>0) {
+			$errors = array();
+			$errors[0] = "This product contains images and can not be delete";
+			return view('inc.admin.showProducts', array('userId'=>$userId, 'categoryId'=>$categoryId, 'subCategoryId'=>$subCategoryId,'errors'=>$errors));
+		}
+
 		$product->delete();
+		redirect("user/$userId/categories/$categoryId/sub_categories/$subCategoryId");
     }
 }
