+++ app/Http/Controllers/SportsController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
 use App\Sport;
+use App\Product;
 use Auth;
 
 class SportsController extends Controller
@@ -16,14 +17,14 @@ class SportsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index($userId)
     {
 		if(!Auth::check() || Auth::user()->role!="root") {
 			abort(404);
 		}
         $sports = Sport::all();
 
-		//lamar a la vista y devolver deportes
+		return view('partials.admin.showSports', array('userId'=>$userId, 'sports'=>$sports));
     }
 
     /**
@@ -121,7 +122,7 @@ class SportsController extends Controller
 		$sport->name = $request->input('name');
 
 		if ($request->hasFile('image')) {
-			if (file_exists($imagePath)) {
+			if (file_exists($sport->imagePath)) {
 				unlink($sport->imagePath);
 			}
 
@@ -155,6 +156,13 @@ class SportsController extends Controller
 
 		if (!$sport = Sport::find($sportId)) {
 			abort(404);
+		}
+
+		$products = Product::where('sportId', $sportId)->get();
+		if (count($products)>0) {
+			$errors = array();
+			$errors[0] = "There are products with this sport and can not be deleted";
+			return view('inc.admin.showSports', array('userId'=>$userId, 'sportId'=>$sportId,'errors'=>$errors));
 		}
 
 		$sport->delete();
